Foi removido a classe Lote, mantendo o numero do lote na classe Vacina.

+++ src/main/java/br/com/municipio/vacinas/vacinas_api/service/VacinaService.java
@@ -10,15 +10,12 @@ import com.mongodb.DuplicateKeyException;
 import lombok.RequiredArgsConstructor;
 
 import br.com.municipio.vacinas.vacinas_api.repository.LocalRepository;
-import br.com.municipio.vacinas.vacinas_api.repository.LoteRepository;
 
 import br.com.municipio.vacinas.vacinas_api.model.Vacina;
 import br.com.municipio.vacinas.vacinas_api.mapper.VacinaMapper;
 import br.com.municipio.vacinas.vacinas_api.dto.VacinaRequestDTO;
 import br.com.municipio.vacinas.vacinas_api.dto.VacinaResponseDTO;
 import br.com.municipio.vacinas.vacinas_api.repository.VacinaRepository;
-
-import br.com.municipio.vacinas.vacinas_api.model.Lote;
 
 /*
     * A classe VacinaService foi implementada para fornecer os serviços relacionados às vacinas, como cadastro, edição, exclusão e busca de vacinas. 
@@ -31,7 +28,6 @@ public class VacinaService {
 
     private final VacinaRepository repository;
     private final LocalRepository localRepository;
-    private final LoteRepository loteRepository;
     private final VacinaMapper mapper;
 
     /*
@@ -55,40 +51,6 @@ public class VacinaService {
                 throw new RuntimeException("Posto de Saúde não encontrado! Por favor, cadastre o Posto de Saúde antes de cadastrar a vacina.");
             }
 
-            /*
-                O bloco de código dentro do if foi implementado para lidar com a situação em que o lote da vacina não existe. 
-                Ele cria um novo lote com as informações fornecidas na requisição, incluindo o número do lote, o fabricante e a vacina associada. 
-                Em seguida, o novo lote é salvo no repositório. 
-                Se o lote já existir, o método atualiza as informações do lote para incluir a nova vacina e fabricante associados a ele, se necessário, e salva as alterações no repositório.
-            */
-            if (!loteRepository.existsByNumeroLote(request.getLote())) {
-                
-                Lote lote = new Lote();
-                lote.setNumeroLote(request.getLote());
-                lote.setFabricante(Arrays.asList(request.getFabricante()));
-                lote.setVacinasAssociadas(Arrays.asList(request.getNome()));
-                loteRepository.save(lote);  
-            } else {
-                /*
-                O bloco de código dentro do else foi implementado para lidar com a situação em que o lote da vacina já existe. 
-                Ele recupera o lote existente do repositório e atualiza as informações do lote para incluir a nova vacina e fabricante associados a ele, se necessário. 
-                Em seguida, as alterações no lote são salvas no repositório.
-                */
-                Lote lote = loteRepository.findByNumeroLote(request.getLote()).orElseThrow(null);
-                List<String> vacinasAssociadas = lote.getVacinasAssociadas();
-                List<String> fabricantes = lote.getFabricante();
-                if (!vacinasAssociadas.contains(request.getNome())) {
-                    vacinasAssociadas.add(request.getNome());
-                    lote.setVacinasAssociadas(vacinasAssociadas);
-                    
-                }
-                if (!fabricantes.contains(request.getFabricante())) {
-                    fabricantes.add(request.getFabricante());
-                    lote.setFabricante(fabricantes);
-                }
-                loteRepository.save(lote);
-            }
-
             return mapper.toDTO(repository.save(mapper.toEntity(request)));
 
         } catch (DuplicateKeyException e) {
@@ -98,54 +60,14 @@ public class VacinaService {
 
     /*
         * O método editarVacina foi implementado para permitir a edição de uma vacina existente. 
-        * Ele recebe os dados atualizados da vacina e o ID da vacina a ser editada. 
-        * O método verifica se a vacina existe e, em seguida, compara o lote atual da vacina com o novo lote fornecido. 
-        * Se os lotes forem iguais, o método atualiza as informações do lote para incluir a nova vacina e fabricante associados a ele, se necessário. 
-        * Se os lotes forem diferentes, o método salva o novo lote e atualiza as informações do lote antigo para remover a associação com a vacina editada, se necessário. 
+        * Ele recebe os dados atualizados da vacina e o ID da vacina a ser editada.
         * Por fim, a vacina editada é salva no repositório e retornada como resposta.
     */
     public VacinaResponseDTO editarVacina(VacinaRequestDTO request, String id) {
-        Vacina vacina = repository.findById(id).orElseThrow(
-            () -> new RuntimeException("Vacina não encontrada!")
-        );
-        String loteAtual = vacina.getLote();
-        String loteNovo = request.getLote();
-        
-        Lote lote = loteRepository.findByNumeroLote(loteNovo).orElseGet(
-            () -> new Lote(null, loteNovo, Arrays.asList(request.getFabricante()), Arrays.asList(request.getNome()))
-        );
+        Vacina vacina = mapper.toEntity(request);
 
-        if (loteAtual.equals(loteNovo)){
-            // Se os lotes forem iguais, o método atualiza as informações do lote para incluir a nova vacina e fabricante associados a ele, se necessário.
-            if (atualizaFabricante(lote, loteAtual, request, vacina)){
-                if (!lote.getFabricante().contains(request.getFabricante()))
-                    lote.getFabricante().add(request.getFabricante());
-            }
-            if (atualizarVacina(lote, loteAtual, request, vacina)){
-                if (!lote.getVacinasAssociadas().contains(request.getNome()))
-                    lote.getVacinasAssociadas().add(request.getNome());
-            }
-
-            loteRepository.save(lote);
-
-        } else {
-            /*
-                O bloco de código dentro do else foi implementado para lidar com a situação em que o lote da vacina está sendo alterado. 
-                Ele salva o novo lote no repositório e, em seguida, atualiza as informações do lote antigo para remover a associação com a vacina editada, se necessário. 
-                Por fim, ele salva as alterações no repositório.
-            */
-            loteRepository.save(lote);
-            Lote l = loteRepository.findByNumeroLote(loteAtual).orElse(null);
-            atualizaFabricante(l, loteAtual, request, vacina);
-            atualizarVacina(l, loteAtual, request, vacina);
-            loteRepository.save(l);
-            
-            if (repository.findByLote(loteAtual).isEmpty()){
-                loteRepository.deleteByNumeroLote(loteAtual);
-            }
-        }
-        vacina = mapper.toEntity(request);
         vacina.setId(id);
+
         return mapper.toDTO(repository.save(vacina));
     }
 
@@ -153,62 +75,12 @@ public class VacinaService {
 
     /*
         * O método excluirVacinaPorId foi implementado para excluir uma vacina com base no seu ID. 
-        * Ele verifica se a vacina existe e, em seguida, remove a vacina do repositório. 
-        * Além disso, ele também verifica se o lote associado à vacina possui outras vacinas ou fabricantes associados. 
-        * Se não houver mais vacinas ou fabricantes associados ao lote, o lote é excluído do repositório.
+        * Ele verifica se a vacina existe e, em seguida, remove a vacina do repositório.
     */
     public void excluirVacinaPorId(String id) {
-        Vacina vacina = repository.findById(id).orElseThrow(
-            () -> new RuntimeException("Vacina não encontrada!")
-        );
-
-        Lote lote = loteRepository.findByNumeroLote(vacina.getLote()).orElseThrow(
-            () -> new RuntimeException("Lote não encontrado!")
-        );
-
-        List<Vacina> vacinas = repository.findByLote(vacina.getLote());
-        boolean flag = false;
-        /* 
-            O loop for foi implementado para verificar se existem outras vacinas associadas ao mesmo fabricante e lote da vacina que está sendo excluída. 
-            Ele percorre a lista de vacinas e verifica se há alguma vacina com o mesmo fabricante e lote, mas com um ID diferente da vacina que está sendo excluída. 
-            Se encontrar uma vacina com essas características, a variável flag é definida como true e o loop é interrompido. 
-            Caso contrário, a variável flag permanece como false, indicando que não há outras vacinas associadas ao mesmo fabricante e lote.
-        */
-        for (Vacina v : vacinas) {
-            if (!v.getId().equals(vacina.getId()) && v.getFabricante().equals(vacina.getFabricante())) {
-                flag = true;
-                break;
-            }
+        if (!repository.existsById(id)) {
+            throw new RuntimeException("Vacina não encontrada!");
         }
-
-        if (!flag){// Se não houver outras vacinas associadas ao mesmo fabricante e lote, o fabricante é removido da lista de fabricantes do lote.
-            lote.getFabricante().remove(vacina.getFabricante());
-        }
-
-        flag = false;
-        /* 
-            O loop for foi implementado para verificar se existem outras vacinas associadas ao mesmo lote e nome da vacina que está sendo excluída. 
-            Ele percorre a lista de vacinas e verifica se há alguma vacina com o mesmo lote e nome, mas com um ID diferente da vacina que está sendo excluída. 
-            Se encontrar uma vacina com essas características, a variável flag é definida como true e o loop é interrompido. 
-            Caso contrário, a variável flag permanece como false, indicando que não há outras vacinas associadas ao mesmo lote e nome.
-        */
-        for (Vacina v : vacinas) {
-            if (!v.getId().equals(vacina.getId()) && v.getNome().equals(vacina.getNome())){
-                flag = true;
-                break;
-            }
-        }
-
-        if (!flag){// Se não houver outras vacinas associadas ao mesmo lote e nome, o nome da vacina é removido da lista de vacinas associadas do lote.
-            lote.getVacinasAssociadas().remove(vacina.getNome());
-        }
-
-        if (lote.getFabricante().isEmpty() && lote.getVacinasAssociadas().isEmpty()) {
-            loteRepository.deleteById(lote.getId());
-        } else {
-            loteRepository.save(lote);
-        }
-
         repository.deleteById(id);
     }
 
@@ -238,51 +110,6 @@ public class VacinaService {
     */
     public List<VacinaResponseDTO> buscarPorNome(String nome){
         return mapper.toDTOList(repository.findByNome(nome));
-    }
-
-    /*
-        * O método atualizaFabricante foi implementado para atualizar a lista de fabricantes associados a um lote.
-    */
-    private boolean atualizaFabricante(Lote lote, String loteAtual, VacinaRequestDTO request, Vacina vacina) {
-        if (!lote.getFabricante().contains(request.getFabricante()) || !request.getFabricante().equals(vacina.getFabricante())){
-            List<Vacina> vacinasList = repository.findAllByFabricante(vacina.getFabricante());
-            boolean flag = false;
-                
-            for (Vacina v : vacinasList) {
-                if (v.getLote().equals(loteAtual) && !v.getId().equals(vacina.getId())){
-                    flag = true;
-                }
-            }
-                
-            if (!flag){
-                lote.getFabricante().remove(vacina.getFabricante());
-            }
-            return true;
-        }
-        return false;
-    }
-
-    /*
-        * O método atualizarVacina foi implementado para atualizar a lista de vacinas associadas a um lote.
-    */
-    private boolean atualizarVacina(Lote lote, String loteAtual, VacinaRequestDTO request, Vacina vacina) {
-        if (!lote.getVacinasAssociadas().contains(request.getNome())|| !request.getNome().equals(vacina.getNome())){
-            List<Vacina> vacinasList = repository.findByNome(vacina.getNome());
-            boolean flag = false;
-
-            for (Vacina v : vacinasList) {
-                if (v.getLote().equals(loteAtual) && !v.getId().equals(vacina.getId())) {
-                    flag = true;
-                    break;
-                }
-            }
-
-            if (!flag){
-                lote.getVacinasAssociadas().remove(vacina.getNome());
-            }
-            return true;        
-        }
-        return false;
     }
 
 }
